showdir a funcionar
pode ainda ser melhorado falta mandar para o cliente a resposta
